pyproject.toml
[tool.poetry]
name = "pydatacolor"
version = "0.1.0"
description = "Connection to Datacolor"
authors = ["Humphrey Drummond <hum3@drummond.info>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pyusb = "^1.2.1"
libusb = "^1.0.26b5"
textual = "^0.22.3"
fastapi = "^0.95.1"
jinja2 = "^3.1.2"
uvicorn = "^0.22.0"
lofigui = "^0.2.3"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^3.6.3"
black = "^23.3.0"
scipy = "^1.10.1"
matplotlib = "^3.7.1"
colour-science = "^0.4.2"
pandas = "^2.0.1"
drawsvg = "^2.1.1"
scikit-learn = "^1.2.2"
colormath = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
